requirements.txt
aiosmtplib==2.0.2
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.0.1
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
cryptography==42.0.2
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-mail==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
orjson==3.9.12
passlib==1.7.4
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.0
pydantic-core==2.16.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
typing-extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
